chore(event): move constants var to constants file
