+++ pkg/controllers/matrixone/constants.go
@@ -1,0 +1,35 @@
+// Copyright 2021 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package matrixone
+
+type matrixoneResourceType string
+type matrixoneActionType string
+type matrixoneActionRes string
+
+const (
+	matrixoneDeployment        matrixoneResourceType = "MatrixoneDeployment"
+	matrixoneStatefulset       matrixoneResourceType = "MatrixoneStatefulset"
+	matrixoneService           matrixoneResourceType = "MatrixoneService"
+	matrixoneCreate            matrixoneActionType   = "MatrixoneCreate"
+	matrixonePatch             matrixoneActionType   = "MatrixonePatch"
+	matrixoneUpdate            matrixoneActionType   = "MatrixoneUpdate"
+	matrixoneDelete            matrixoneActionType   = "MatrixoneDelete"
+	matrixoneRollingDeployWait matrixoneActionType   = "MatrixoneRollydeployWait"
+	matrixoneObjectGet         matrixoneActionType   = "MatrixoneObjectGet"
+	matrixoneList              matrixoneActionType   = "matrixoneList"
+
+	matrixoneActionFailed    matrixoneActionRes = "MatrixoneActionFailed"
+	matrixoneActionSuccessed matrixoneActionRes = "MatrixoneActionSuccessed"
+)
